% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiTxDb.R
\docType{package}
\name{EpiTxDb-package}
\alias{EpiTxDb-package}
\title{EpiTxDb: Storing and accessing epitranscriptomic information using the AnnotationDbi interface}
\description{
EpiTxDb facilitates the storage of epitranscriptomic information. More specifically, it can keep track of modification identity, position, the enzyme for introducing it on the RNA, a specifier which determines the position on the RNA to be modified and the literature references each modification is associated with.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FelixErnst/EpiTxDb}
  \item Report bugs at \url{https://github.com/FelixErnst/EpiTxDb/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

}
\keyword{internal}
