% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTxModDbFromGRanges.R,
%   R/makeTxModDbFromtRNAdb.R
\name{makeTxModDbfromtRNAdb}
\alias{makeTxModDbfromtRNAdb}
\alias{makeTxModDbfromGRanges}
\alias{listAvailableOrganismsfromtRNAdb}
\title{makeTxModDbfromtRNAdb}
\usage{
makeTxModDbfromGRanges(gr, metadata = NULL, reassign.ids = FALSE)

makeTxModDbfromtRNAdb(organism, txdb, genome, metadata = NULL)

listAvailableOrganismsfromtRNAdb()
}
\description{
title

title
}
