% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDb.R
\name{makeEpiTxDb}
\alias{makeEpiTxDb}
\title{makeEpiTxDb}
\usage{
makeEpiTxDb(
  modifications,
  reactions = NULL,
  specifiers = NULL,
  references = NULL,
  metadata = NULL,
  reassign.ids = FALSE
)
}
\description{
title
}
