% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDb.R
\name{makeEpiTxDb}
\alias{makeEpiTxDb}
\title{Creating a \code{EpiTxDb} from user supplied annotations as
\code{data.frame}s}
\usage{
makeEpiTxDb(
  modifications,
  reactions = NULL,
  specifiers = NULL,
  references = NULL,
  metadata = NULL,
  reassign.ids = FALSE
)
}
\arguments{
\item{modifications}{A \code{data.frame} containg the following columns:
\itemize{
  \item{\code{mod_id} }
  \item{\code{mod_type} }
  \item{\code{mod_name} }
  \item{\code{mod_start} }
  \item{\code{mod_end} }
  \item{\code{sn_id} }
  \item{\code{sn_name} }
}
The first three are mandatory, whereas one of the last has to be set.
\code{sn_id} will be generated from the other the last two. The last three
columns must be consistent in meaning.}

\item{reactions}{An optional \code{data.frame} containg the following
columns:
\itemize{
  \item{\code{rx_genename} }
  \item{\code{rx_rank} }
  \item{\code{rx_ensembl} }
  \item{\code{rx_ensembltrans} }
  \item{\code{rx_entrezid} }
}
(default: \code{reactions = NULL})}

\item{specifiers}{An optional \code{data.frame} containg the following
columns:
  \itemize{
  \item{\code{spec_genename} }
  \item{\code{spec_rank} }
  \item{\code{spec_ensembl} }
  \item{\code{spec_ensembltrans} }
  \item{\code{spec_entrezid} }
}
(default: \code{specifiers = NULL})}

\item{references}{An optional \code{data.frame} containg the following
columns:
\itemize{
  \item{\code{ref_type} }
  \item{\code{ref} }
}
(default: \code{references = NULL})}

\item{metadata}{An optional \code{data.frame} containg the following columns:
\itemize{
  \item{\code{name}: }{a character value used as name}
  \item{\code{value}: }{a character value}
}
This dataframe will be returned
by \code{metadata()} (default: \code{metadata = NULL})}

\item{reassign.ids}{\code{TRUE} or \code{FALSE} Controls how internal
\code{mod_id}s should be assigned. If \code{reassign.ids} is \code{FALSE}
(the default) and if the ids are supplied, then they are used as the
internal ids, otherwise the internal ids are assigned in a way that is
compatible with the order defined by ordering the modifications first by
chromosome, then by strand, then by start, and finally by end.}
}
\description{
\code{makeEpiTxDb} is a low-level constructor for creating a
\code{\link[=EpiTxDb-class]{EpiTxDb}} object from user supplied annotations.

This functions probably will typically not be used by users.
}
\examples{

}
\seealso{
\itemize{
  \item{\code{\link[=makeEpiTxDbFromGRanges]{makeEpiTxDbFromGRanges}} for
    creating a \code{EpiTxDb} object from a
  \code{\link[GenomicRanges:GRanges-class]{GRanges}} object and it's metadata
    columns}
  \item{\code{\link[=makeEpiTxDbFromRMBase]{makeEpiTxDbFromRMBase}} for
    creating a \code{EpiTxDb} object from RMBase online resources}
  \item{\code{\link[=makeEpiTxDbFromtRNAdb]{makeEpiTxDbFromtRNAdb}} for
    creating a \code{EpiTxDb} object from tRNAdb online resources}
}
}
