% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDbFromRMBase.R
\name{makeEpiTxDbfromRMBase}
\alias{makeEpiTxDbfromRMBase}
\alias{downloadRMBaseFiles}
\alias{getRMBaseData}
\alias{makeEpiTxDbfromRMBaseFiles}
\alias{listAvailableOrganismsFromRMBase}
\alias{listAvailableGenomesFromRMBase}
\alias{listAvailableModFromRMBase}
\title{makeEpiTxDbfromRMBase}
\usage{
downloadRMBaseFiles(organism, genome, type)

makeEpiTxDbfromRMBase(
  organism,
  genome,
  type,
  tx,
  sequences = NULL,
  metadata = NULL,
  reassign.ids = FALSE
)

getRMBaseData(files, seqlevels = NA)

makeEpiTxDbfromRMBaseFiles(
  files,
  tx,
  sequences = NULL,
  metadata = NULL,
  reassign.ids = FALSE
)

listAvailableOrganismsFromRMBase()

listAvailableGenomesFromRMBase(organism)

listAvailableModFromRMBase(organism, genome)
}
\description{
title
}
