% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDbFromRMBase.R
\name{makeEpiTxDbfromRMBase}
\alias{makeEpiTxDbfromRMBase}
\alias{makeEpiTxDbfromRMBaseFile}
\alias{listAvailableOrganismsFromRMBase}
\alias{listAvailableGenomesFromRMBase}
\alias{listAvailableModFromRMBase}
\title{makeEpiTxDbfromRMBase}
\usage{
makeEpiTxDbfromRMBase(
  organism,
  genome,
  type,
  tx,
  metadata = NULL,
  reassign.ids = FALSE
)

makeEpiTxDbfromRMBaseFile(files, tx, metadata = NULL, reassign.ids = FALSE)

listAvailableOrganismsFromRMBase()

listAvailableGenomesFromRMBase(organism)

listAvailableModFromRMBase(organism, genome)
}
\description{
title
}
