% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDbFromtRNAdb.R
\name{makeEpiTxDbfromtRNAdb}
\alias{makeEpiTxDbfromtRNAdb}
\alias{listAvailableOrganismsfromtRNAdb}
\title{makeEpiTxDbfromtRNAdb}
\usage{
makeEpiTxDbfromtRNAdb(organism, txdb, sequences, metadata = NULL)

listAvailableOrganismsfromtRNAdb()
}
\description{
title
}
