% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpiTxDbFromtRNAdb.R
\name{makeEpiTxDbFromtRNAdb}
\alias{makeEpiTxDbFromtRNAdb}
\alias{listAvailableOrganismsFromtRNAdb}
\title{makeEpiTxDbFromtRNAdb}
\usage{
makeEpiTxDbFromtRNAdb(organism, tx = NULL, sequences = NULL, metadata = NULL)

listAvailableOrganismsFromtRNAdb()
}
\description{
title
}
