% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R, R/modificationsBy.R
\name{modifications}
\alias{modifications}
\alias{modificationsBy}
\alias{modifications,TxModDb-method}
\alias{modificationsByTranscript,TxModDb-method}
\title{modifications}
\usage{
\S4method{modifications}{TxModDb}(
  x,
  column = c("mod_id", "mod_type", "mod_name"),
  filter = NULL,
  use.names = FALSE
)

\S4method{modificationsByTranscript}{TxModDb}(x, use.names = FALSE)
}
\arguments{
\item{x}{a \code{\link[=TxModDb-class]{TxModDb}}}

\item{filter}{XXX}

\item{columns}{XXX}
}
\description{
title
}
\examples{
\dontrun{
library(TxModDb.Hsapiens.hg38)
modifications(TxModDb.Hsapiens.hg38))
}
}
