% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R, R/modificationsBy.R
\name{modifications}
\alias{modifications}
\alias{modificationsBy}
\alias{modifications,EpiTxDb-method}
\alias{modificationsBy,EpiTxDb-method}
\alias{modifiedSeqsByTranscript,EpiTxDb,DNAStringSet-method}
\title{modifications}
\usage{
\S4method{modifications}{EpiTxDb}(
  x,
  column = c("mod_id", "mod_type", "mod_name"),
  filter = NULL,
  use.names = FALSE
)

\S4method{modificationsBy}{EpiTxDb}(x, by = c("transcript", "type", "reaction", "specifier"))

\S4method{modifiedSeqsByTranscript}{EpiTxDb,DNAStringSet}(x, sequences)
}
\arguments{
\item{x}{a \code{\link[=EpiTxDb-class]{EpiTxDb}}}

\item{filter}{XXX}

\item{columns}{XXX}
}
\description{
title
}
\examples{
\dontrun{
library(EpiTxDb.Hsapiens.hg38)
modifications(EpiTxDb.Hsapiens.hg38))
}
}
