% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/shiftGenomicToTranscript.R
\name{shiftTranscriptToGenomic}
\alias{shiftTranscriptToGenomic}
\alias{shiftGenomicToTranscript}
\alias{shiftTranscriptToGenomic,GRanges,GRangesList-method}
\alias{shiftTranscriptToGenomic,GRangesList,GRangesList-method}
\alias{shiftGenomicToTranscript,GRanges,GRangesList-method}
\alias{shiftGenomicToTranscript,GRangesList,GRangesList-method}
\title{Shift \code{GRanges} coordinates based on another \code{GRanges}
object}
\usage{
shiftTranscriptToGenomic(subject, tx)

shiftGenomicToTranscript(subject, tx)

\S4method{shiftTranscriptToGenomic}{GRanges,GRangesList}(subject, tx)

\S4method{shiftTranscriptToGenomic}{GRangesList,GRangesList}(subject, tx)

\S4method{shiftGenomicToTranscript}{GRanges,GRangesList}(subject, tx)

\S4method{shiftGenomicToTranscript}{GRangesList,GRangesList}(subject, tx)
}
\arguments{
\item{subject}{a \code{\link[GenomicRanges:GRanges-class]{GRanges}} or
\code{\link[GenomicRanges:GRangesList-class]{GRangesList}} object}

\item{tx}{a \code{\link[GenomicRanges:GRangesList-class]{GRangesList}} object
with a \code{type} metadata column. In addition the metadata \code{mcols(tx)}
must contain at least two columns \code{gene_name} and \code{transcript_name}
defining the relation of gene and transcript names.}
}
\description{
\code{shiftGenomicToTranscript} shifts positions of a
\code{\link[GenomicRanges:GRanges-class]{GRanges}}
object based on coordinates of another \code{GRanges} object. The most common
application is to shift genomic coordinates to transcript coordinates, which
is reflected in the name. \code{shiftTranscriptToGenomic} implements the
reverse operation.
}
\examples{
\dontrun{
library(GenomicRanges)
# Construct some example data
gr1 <- GRanges("chr2", IRanges(3, 6),
               strand="+", gene_name=c("c_g"))
gr2 <- GRanges(c("chr1", "chr1"), IRanges(c(7,13), width=3),
               strand=c("+", "-"), gene_name=c("a_g","b_g"))
gr3 <- GRanges(c("chr1", "chr2"), IRanges(c(1, 4), c(3, 9)),
               strand=c("-", "-"), gene_name=c("a_g","c_g"))
grl <- GRangesList(gr1=gr1, gr2=gr2, gr3=gr3)
tx1 <- GRanges("chr1", IRanges(1, 10),
               strand="+", type="exon")
tx2 <- GRanges("chr1", IRanges(10, 20),
               strand="+", type="exon")
tx3 <- GRanges("chr2", IRanges(1, 10),
               strand="-", type="exon")
tx <- GRangesList(a=tx1, b=tx2, c=tx3)
mcols(tx) <- DataFrame(gene_name = c("a_g","b_g","c_g"),
                       tx_name = c("a","b","c"))

# shift to transcript coordinates
shifted_grl <- shiftGenomicToTranscript(grl,tx)
# ... and back
shifted_grl2 <- shiftTranscriptToGenomic(shifted_grl,tx)
# comparison of ranges work. However the seqlevels differ
ranges(shifted_grl2) == ranges(grl)
}
}
